requirements.txt
fastmcp>=1.0.0
docker>=7.0.0
psutil>=5.9.0
aiohttp>=3.9.0
cryptography>=42.0.0
pydantic>=2.0.0
pyyaml>=6.0.0
uvicorn>=0.25.0
python-multipart>=0.0.6
requests>=2.31.0
dnspython>=2.4.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.7.0
flake8>=6.0.0
pre-commit>=3.5.0
